Move date constants to date util.
