+++ src/components/Modal.css
@@ -41,7 +41,7 @@
   .modal-content {
     position: static;
     border-radius: 0.25rem;
-    max-width: 40rem;
+    max-width: 35rem;
   }
 }
